# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LedgerGuard AI Treasury OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List connected wallets
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
    post:
      operationId: connectWallet
      tags: [wallets]
      summary: Connect a wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletInput"
      responses:
        "201":
          description: Wallet connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallets/{id}:
    delete:
      operationId: disconnectWallet
      tags: [wallets]
      summary: Disconnect a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disconnected
  /wallets/{id}/balance:
    get:
      operationId: getWalletBalance
      tags: [wallets]
      summary: Get live balance for a wallet from Solana
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"

  # ── Transactions ──────────────────────────────────────────────────────────────
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update transaction status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/simulate:
    post:
      operationId: simulateTransaction
      tags: [transactions]
      summary: Simulate a transaction on Solana
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
  /transactions/{id}/broadcast:
    post:
      operationId: broadcastTransaction
      tags: [transactions]
      summary: Broadcast a signed transaction to Solana
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Broadcast result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResult"
  /transactions/{id}/risk:
    get:
      operationId: getTransactionRisk
      tags: [transactions]
      summary: Get risk score for a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAssessment"

  # ── Policies ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Wallet:
      type: object
      required: [id, address, network, createdAt]
      properties:
        id:
          type: integer
        address:
          type: string
        network:
          type: string
        label:
          type: ["string", "null"]
        createdAt:
          type: string
    WalletInput:
      type: object
      required: [address, network]
      properties:
        address:
          type: string
        network:
          type: string
        label:
          type: string
    WalletBalance:
      type: object
      required: [address, balanceLamports, balanceSol, network]
      properties:
        address:
          type: string
        balanceLamports:
          type: number
        balanceSol:
          type: number
        network:
          type: string

    # ── Transactions ───────────────────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, amount, recipient, token, status, riskScore, network, createdAt]
      properties:
        id:
          type: integer
        signature:
          type: ["string", "null"]
        amount:
          type: number
        recipient:
          type: string
        token:
          type: string
        memo:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, signed, broadcast, confirmed, failed]
        riskScore:
          type: ["integer", "null"]
        riskLevel:
          type: ["string", "null"]
        network:
          type: string
        fromWalletAddress:
          type: ["string", "null"]
        aiProposed:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    TransactionInput:
      type: object
      required: [amount, recipient, token, network]
      properties:
        amount:
          type: number
        recipient:
          type: string
        token:
          type: string
        memo:
          type: string
        network:
          type: string
        fromWalletAddress:
          type: string
        aiProposed:
          type: boolean
    TransactionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, signed, broadcast, confirmed, failed]
        signature:
          type: string
        riskScore:
          type: integer
        riskLevel:
          type: string
    SimulationResult:
      type: object
      required: [success, estimatedFee, logs]
      properties:
        success:
          type: boolean
        estimatedFee:
          type: number
        logs:
          type: array
          items:
            type: string
        error:
          type: ["string", "null"]
    BroadcastInput:
      type: object
      required: [signedTransaction]
      properties:
        signedTransaction:
          type: string
    BroadcastResult:
      type: object
      required: [signature, success]
      properties:
        signature:
          type: string
        success:
          type: boolean
        error:
          type: ["string", "null"]
    RiskAssessment:
      type: object
      required: [riskScore, level, reasons]
      properties:
        riskScore:
          type: integer
        level:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        reasons:
          type: array
          items:
            type: string

    # ── Policies ───────────────────────────────────────────────────────────────
...